Introduce a key value `${filename}`
Now this value will always replaced with the current files name.
When uploading files. This allows us to manipulate the path to
prevent duplicates when uploading photos.

Worldly now prefixes each file name with a folder prefixing the
number of the photo it is in the batch:

0/image.jpg
1/image.jpg
2/image.jpg

This ensures that there will never be a file name conflict.